ASoC: SOF: intel: do not init i915 codec if codec is masked
The SOF HDA driver goes through a 60s timeout if the i915 driver
is missing or not fully functional in early silicon enablement.
When we already have the information through the codec_mask that
the display codec is not reported, we can avoid the i915
initialization altogether and skip the timeout.

Also add check mask function in exit to make sure power steps are
paired in error handling or exit process.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>